requirements.txt
certifi==2023.11.17
influxdb-client==1.39.0
paho-mqtt==1.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
reactivex==4.0.4
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
